requirements.txt
flask
mysql-connector-python 
Flask-Login
Flask-WTF WTForms
Flask-Mail
email_validator
pip
SQLAlchemy